Refactor component imports in CourseOverview and CourseList pages
