+++ src/components/atoms/vCourseAvatar.unit.ts
