+++ src/components/atoms/vCourseAvatar.vue
