+++ src/components/atoms/vCourseEmptyAvatar.vue
